refactor(node): replace std/path references from std/node
