+++ node/testdata/child_process_unref.js
@@ -1,5 +1,5 @@
 import cp from "../child_process.ts";
-import * as path from "../../path/mod.ts";
+import * as path from "../path.ts";
 
 const script = path.join(
   path.dirname(path.fromFileUrl(import.meta.url)),
